requirement.txt
Django==5.0.7
djangorestframework==3.15.2
Markdown==3.6
django-filter==24.2